Added WASP RNA-seq remapping
I will have to think about how to refactor this. The remapping command
is the same for ATAC and RNA, the only difference is that actual call to
STAR (which has different parameters for the different data types). I
could put the remapping command in general and pass an option for the
data type and choose the appropriate alignment method.