Updates for new cluster
Many changes for new cluster. I'm going to more radically change the
structure of this package to better use the resources on the new
cluster. Namely, I'm going to have one master method that runs the
RNA-seq pipeline. This method will spawn several job scripts and a
master script to submit them.