Updates for allelic imbalance/ASE
I've updated the RNA-seq pipeline so that ASE can be calculated when the
VCF file and the RNA-seq have different chromosome naming. I need to
propagate this to the ATAC-seq pipeline. I'm also going to update so
that the pipeline can accept multiple VCF files (like if the VCFs are
split up by chromosome).